relations, swagger

+++ SPRING_BOOT/lesson 33/Student/src/main/java/com/ltc/student/repository/TeacherRepository.java
@@ -3,5 +3,5 @@ package com.ltc.student.repository;
 import com.ltc.student.entity.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Integer> {
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
 }
